API for ciphertext header versioning.
Define simple API for extensible ciphertext header versioning.
