+++ pom.xml
@@ -4,7 +4,7 @@
   <groupId>org.cryptacular</groupId>
   <artifactId>cryptacular</artifactId>
   <packaging>jar</packaging>
-  <version>1.2.5-SNAPSHOT</version>
+  <version>1.3.0-SNAPSHOT</version>
   <name>Cryptacular Library</name>
   <description>The spectacular complement to the Bouncy Castle crypto API for Java.</description>
   <url>http://www.cryptacular.org</url>
