+++ src/main/java/org/cryptacular/util/CipherUtil.java
@@ -4,7 +4,6 @@ package org.cryptacular.util;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.function.Function;
 import javax.crypto.SecretKey;
 import org.bouncycastle.crypto.BlockCipher;
 import org.bouncycastle.crypto.modes.AEADBlockCipher;
@@ -14,10 +13,12 @@ import org.bouncycastle.crypto.params.AEADParameters;
 import org.bouncycastle.crypto.params.KeyParameter;
 import org.bouncycastle.crypto.params.ParametersWithIV;
 import org.cryptacular.CiphertextHeader;
-import org.cryptacular.CiphertextHeaderV2;
+import org.cryptacular.CiphertextHeaderFactory;
 import org.cryptacular.CryptoException;
 import org.cryptacular.EncodingException;
 import org.cryptacular.StreamException;
+import org.cryptacular.V1CiphertextHeader;
+import org.cryptacular.V2CiphertextHeader;
 import org.cryptacular.adapter.AEADBlockCipherAdapter;
 import org.cryptacular.adapter.BlockCipherAdapter;
 import org.cryptacular.adapter.BufferedBlockCipherAdapter;
@@ -39,7 +40,7 @@ public final class CipherUtil
 
 
   /**
-   * Encrypts data using an AEAD cipher. A {@link CiphertextHeaderV2} is prepended to the resulting ciphertext and
+   * Encrypts data using an AEAD cipher. A {@link V2CiphertextHeader} is prepended to the resulting ciphertext and
    * used as AAD (Additional Authenticated Data) passed to the AEAD cipher.
    *
    * @param  cipher  AEAD cipher.
@@ -47,7 +48,7 @@ public final class CipherUtil
    * @param  nonce  Nonce generator.
    * @param  data  Plaintext data to be encrypted.
    *
-   * @return  Concatenation of encoded {@link CiphertextHeaderV2} and encrypted data that completely fills the returned
+   * @return  Concatenation of encoded {@link V2CiphertextHeader} and encrypted data that completely fills the returned
    *          byte array.
    *
    * @throws  CryptoException  on encryption errors.
@@ -56,21 +57,21 @@ public final class CipherUtil
     throws CryptoException
   {
     final byte[] iv = nonce.generate();
-    final byte[] header = new CiphertextHeaderV2(iv, "1").encode(key);
+    final byte[] header = new V2CiphertextHeader(iv, "1").encode(key);
     cipher.init(true, new AEADParameters(new KeyParameter(key.getEncoded()), MAC_SIZE_BITS, iv, header));
     return encrypt(new AEADBlockCipherAdapter(cipher), header, data);
   }
 
 
   /**
-   * Encrypts data using an AEAD cipher. A {@link CiphertextHeaderV2} is prepended to the resulting ciphertext and used
+   * Encrypts data using an AEAD cipher. A {@link V2CiphertextHeader} is prepended to the resulting ciphertext and used
    * as AAD (Additional Authenticated Data) passed to the AEAD cipher.
    *
    * @param  cipher  AEAD cipher.
    * @param  key  Encryption key.
    * @param  nonce  Nonce generator.
    * @param  input  Input stream containing plaintext data.
-   * @param  output  Output stream that receives a {@link CiphertextHeaderV2} followed by ciphertext data produced by
+   * @param  output  Output stream that receives a {@link V2CiphertextHeader} followed by ciphertext data produced by
    *                 the AEAD cipher in encryption mode.
    *
    * @throws  CryptoException  on encryption errors.
@@ -85,7 +86,7 @@ public final class CipherUtil
     throws CryptoException, StreamException
   {
     final byte[] iv = nonce.generate();
-    final byte[] header = new CiphertextHeaderV2(iv, "1").encode(key);
+    final byte[] header = new V2CiphertextHeader(iv, "1").encode(key);
     cipher.init(true, new AEADParameters(new KeyParameter(key.getEncoded()), MAC_SIZE_BITS, iv, header));
     writeHeader(header, output);
     process(new AEADBlockCipherAdapter(cipher), input, output);
@@ -97,7 +98,7 @@ public final class CipherUtil
    *
    * @param  cipher  AEAD cipher.
    * @param  key  Encryption key.
-   * @param  data  Ciphertext data containing a prepended {@link CiphertextHeaderV2}. The header is treated as AAD input
+   * @param  data  Ciphertext data containing a prepended {@link V2CiphertextHeader}. The header is treated as AAD input
    *               to the cipher that is verified during decryption.
    *
    * @return  Decrypted data that completely fills the returned byte array.
@@ -108,7 +109,7 @@ public final class CipherUtil
   public static byte[] decrypt(final AEADBlockCipher cipher, final SecretKey key, final byte[] data)
       throws CryptoException, EncodingException
   {
-    final CiphertextHeader header = decodeHeader(data, String -> key);
+    final CiphertextHeader header = CiphertextHeaderFactory.decode(data, String -> key);
     final byte[] nonce = header.getNonce();
     final byte[] hbytes = header.encode();
     cipher.init(false, new AEADParameters(new KeyParameter(key.getEncoded()), MAC_SIZE_BITS, nonce, hbytes));
@@ -121,7 +122,7 @@ public final class CipherUtil
    *
    * @param  cipher  AEAD cipher.
    * @param  key  Encryption key.
-   * @param  input  Input stream containing a {@link CiphertextHeaderV2} followed by ciphertext data. The header is
+   * @param  input  Input stream containing a {@link V2CiphertextHeader} followed by ciphertext data. The header is
    *                treated as AAD input to the cipher that is verified during decryption.
    * @param  output  Output stream that receives plaintext produced by block cipher in decryption mode.
    *
@@ -136,7 +137,7 @@ public final class CipherUtil
     final OutputStream output)
     throws CryptoException, EncodingException, StreamException
   {
-    final CiphertextHeader header = decodeHeader(input, String -> key);
+    final CiphertextHeader header = CiphertextHeaderFactory.decode(input, String -> key);
     final byte[] nonce = header.getNonce();
     final byte[] hbytes = header.encode();
     cipher.init(false, new AEADParameters(new KeyParameter(key.getEncoded()), MAC_SIZE_BITS, nonce, hbytes));
@@ -145,7 +146,7 @@ public final class CipherUtil
 
 
   /**
-   * Encrypts data using the given block cipher with PKCS5 padding. A {@link CiphertextHeaderV2} is prepended to the
+   * Encrypts data using the given block cipher with PKCS5 padding. A {@link V2CiphertextHeader} is prepended to the
    * resulting ciphertext.
    *
    * @param  cipher  Block cipher.
@@ -154,7 +155,7 @@ public final class CipherUtil
    *                cipher block size.
    * @param  data  Plaintext data to be encrypted.
    *
-   * @return  Concatenation of encoded {@link CiphertextHeaderV2} and encrypted data that completely fills the returned
+   * @return  Concatenation of encoded {@link V2CiphertextHeader} and encrypted data that completely fills the returned
    *          byte array.
    *
    * @throws  CryptoException  on encryption errors.
@@ -163,7 +164,7 @@ public final class CipherUtil
     throws CryptoException
   {
     final byte[] iv = nonce.generate();
-    final byte[] header = new CiphertextHeaderV2(iv, "1").encode(key);
+    final byte[] header = new V2CiphertextHeader(iv, "1").encode(key);
     final PaddedBufferedBlockCipher padded = new PaddedBufferedBlockCipher(cipher, new PKCS7Padding());
     padded.init(true, new ParametersWithIV(new KeyParameter(key.getEncoded()), iv));
     return encrypt(new BufferedBlockCipherAdapter(padded), header, data);
@@ -171,7 +172,7 @@ public final class CipherUtil
 
 
   /**
-   * Encrypts data using the given block cipher with PKCS5 padding. A {@link CiphertextHeader} is prepended to the
+   * Encrypts data using the given block cipher with PKCS5 padding. A {@link V1CiphertextHeader} is prepended to the
    * resulting ciphertext.
    *
    * @param  cipher  Block cipher.
@@ -193,7 +194,7 @@ public final class CipherUtil
     throws CryptoException, StreamException
   {
     final byte[] iv = nonce.generate();
-    final byte[] header = new CiphertextHeaderV2(iv, "1").encode(key);
+    final byte[] header = new V2CiphertextHeader(iv, "1").encode(key);
     final PaddedBufferedBlockCipher padded = new PaddedBufferedBlockCipher(cipher, new PKCS7Padding());
     padded.init(true, new ParametersWithIV(new KeyParameter(key.getEncoded()), iv));
     writeHeader(header, output);
@@ -206,7 +207,7 @@ public final class CipherUtil
    *
    * @param  cipher  Block cipher.
    * @param  key  Encryption key.
-   * @param  data  Ciphertext data containing a prepended {@link CiphertextHeader}.
+   * @param  data  Ciphertext data containing a prepended {@link V1CiphertextHeader}.
    *
    * @return  Decrypted data that completely fills the returned byte array.
    *
@@ -216,7 +217,7 @@ public final class CipherUtil
   public static byte[] decrypt(final BlockCipher cipher, final SecretKey key, final byte[] data)
     throws CryptoException, EncodingException
   {
-    final CiphertextHeader header = decodeHeader(data, String -> key);
+    final CiphertextHeader header = CiphertextHeaderFactory.decode(data, String -> key);
     final PaddedBufferedBlockCipher padded = new PaddedBufferedBlockCipher(cipher, new PKCS7Padding());
     padded.init(false, new ParametersWithIV(new KeyParameter(key.getEncoded()), header.getNonce()));
     return decrypt(new BufferedBlockCipherAdapter(padded), data, header.getLength());
@@ -228,7 +229,7 @@ public final class CipherUtil
    *
    * @param  cipher  Block cipher.
    * @param  key  Encryption key.
-   * @param  input  Input stream containing a {@link CiphertextHeader} followed by ciphertext data.
+   * @param  input  Input stream containing a {@link V1CiphertextHeader} followed by ciphertext data.
    * @param  output  Output stream that receives plaintext produced by block cipher in decryption mode.
    *
    * @throws  CryptoException  on encryption errors.
@@ -242,59 +243,10 @@ public final class CipherUtil
     final OutputStream output)
     throws CryptoException, EncodingException, StreamException
   {
-    final CiphertextHeader header = decodeHeader(input, String -> key);
+    final CiphertextHeader header = CiphertextHeaderFactory.decode(input, String -> key);
     final PaddedBufferedBlockCipher padded = new PaddedBufferedBlockCipher(cipher, new PKCS7Padding());
     padded.init(false, new ParametersWithIV(new KeyParameter(key.getEncoded()), header.getNonce()));
     process(new BufferedBlockCipherAdapter(padded), input, output);
-  }
-
-
-  /**
-   * Decodes the ciphertext header at the start of the given byte array.
-   * Supports both original (deprecated) and v2 formats.
-   *
-   * @param  data  Ciphertext data with prepended header.
-   * @param  keyLookup  Decryption key lookup function.
-   *
-   * @return  Ciphertext header instance.
-   */
-  public static CiphertextHeader decodeHeader(final byte[] data, final Function<String, SecretKey> keyLookup)
-  {
-    try {
-      return CiphertextHeaderV2.decode(data, keyLookup);
-    } catch (EncodingException e) {
-      return CiphertextHeader.decode(data);
-    }
-  }
-
-
-  /**
-   * Decodes the ciphertext header at the start of the given input stream.
-   * Supports both original (deprecated) and v2 formats.
-   *
-   * @param  in  Ciphertext stream that is positioned at the start of the ciphertext header.
-   * @param  keyLookup  Decryption key lookup function.
-   *
-   * @return  Ciphertext header instance.
-   */
-  public static CiphertextHeader decodeHeader(final InputStream in, final Function<String, SecretKey> keyLookup)
-  {
-    CiphertextHeader header;
-    try {
-      // Mark the stream start position so we can try again with old format header
-      if (in.markSupported()) {
-        in.mark(4);
-      }
-      header = CiphertextHeaderV2.decode(in, keyLookup);
-    } catch (EncodingException e) {
-      try {
-        in.reset();
-      } catch (IOException ioe) {
-        throw new StreamException("Stream error trying to process old header format: " + ioe.getMessage());
-      }
-      header = CiphertextHeader.decode(in);
-    }
-    return header;
   }
 
 
@@ -360,7 +312,7 @@ public final class CipherUtil
     final int inSize = 1024;
     final int outSize = cipher.getOutputSize(inSize);
     final byte[] inBuf = new byte[inSize];
-    final byte[] outBuf = new byte[outSize > inSize ? outSize : inSize];
+    final byte[] outBuf = new byte[Math.max(outSize, inSize)];
     int readLen;
     int writeLen;
     try {
