+++ src/test/java/org/cryptacular/V1CiphertextHeaderTest.java
@@ -5,13 +5,14 @@ import java.util.Arrays;
 import org.cryptacular.util.CodecUtil;
 import org.testng.annotations.Test;
 import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertTrue;
 
 /**
- * Unit test for {@link CiphertextHeader}.
+ * Unit test for {@link V1CiphertextHeader}.
  *
  * @author Middleware Services
  */
-public class CiphertextHeaderTest
+public class V1CiphertextHeaderTest
 {
 
   @Test(
@@ -19,7 +20,7 @@ public class CiphertextHeaderTest
       expectedExceptionsMessageRegExp = "Nonce exceeds size limit in bytes.*")
   public void testNonceLimitConstructor()
   {
-    new CiphertextHeader(new byte[256], "key2");
+    new V1CiphertextHeader(new byte[256], "key2");
   }
 
   @Test
@@ -27,10 +28,11 @@ public class CiphertextHeaderTest
   {
     final byte[] nonce = new byte[255];
     Arrays.fill(nonce, (byte) 7);
-    final CiphertextHeader expected = new CiphertextHeader(nonce, "aleph");
+    final V1CiphertextHeader expected = new V1CiphertextHeader(nonce, "aleph");
     final byte[] encoded = expected.encode();
     assertEquals(expected.getLength(), encoded.length);
-    final CiphertextHeader actual = CiphertextHeader.decode(encoded);
+    final CiphertextHeader actual = CiphertextHeaderFactory.decode(encoded, null);
+    assertTrue(actual instanceof V1CiphertextHeader);
     assertEquals(expected.getNonce(), actual.getNonce());
     assertEquals(expected.getKeyName(), actual.getKeyName());
     assertEquals(expected.getLength(), actual.getLength());
@@ -42,7 +44,7 @@ public class CiphertextHeaderTest
   public void testDecodeFailNonceLengthExceeded()
   {
     // https://github.com/vt-middleware/cryptacular/issues/52
-    CiphertextHeader.decode(CodecUtil.hex("000000347ffffffd"));
+    CiphertextHeaderFactory.decode(CodecUtil.hex("000000347ffffffd"), null);
   }
 
   @Test(
@@ -50,6 +52,6 @@ public class CiphertextHeaderTest
       expectedExceptionsMessageRegExp = "Bad ciphertext header: maximum key length exceeded")
   public void testDecodeFailKeyLengthExceeded()
   {
-    CiphertextHeader.decode(CodecUtil.hex("000000F300000004DEADBEEF00FFFFFF"));
+    CiphertextHeaderFactory.decode(CodecUtil.hex("000000F300000004DEADBEEF00FFFFFF"), null);
   }
 }
